cleanup macros
